common: Put `WithId` into a separate module
